+++ waylib/src/server/qtquick/wrenderhelper.cpp
@@ -703,7 +703,7 @@ QSGRendererInterface::GraphicsApi WRenderHelper::probe(qw_backend *testBackend, 
             std::unique_ptr<qw_allocator> alloc(qw_allocator::autocreate(*testBackend, *renderer.get()));
 
             bool hasSupportedFormat = false;
-            for (int formatId = 0; formatId < formats->len; formatId++) {
+            for (size_t formatId = 0; formatId < formats->len; formatId++) {
                 auto *format = &formats->formats[formatId];
 
                 std::unique_ptr<qw_swapchain> swapchain(qw_swapchain::create(*alloc.get(), 1000, 800, format));
